engine/requirements-dev.txt
celery-types==0.17.0
django-filter-stubs==0.1.3
django-stubs[compatible-mypy]==4.2.1
djangorestframework-stubs[compatible-mypy]==3.14.1
mypy==1.3.0
pre-commit==2.15.0
pytest==7.3.1
pytest-django==4.5.2
pytest_factoryboy==2.5.1
types-beautifulsoup4==4.12.0.5
types-PyMySQL==1.0.19.7
types-python-dateutil==2.8.19.13
